requirements.txt
pyserial>=3.5
plyer>=2.0.0